Moved to floats for lat/lon data. I'm not actually looking for more precision than that for this app. Should save a couple GB of RAM on the biggest file entries (Norway, some Canadian provinces)

+++ DatabaseAccess/Support/Records.cs
@@ -7,13 +7,13 @@ using System.Threading.Tasks;
 namespace DatabaseAccess.Support
 {
     //records are new C# 9.0 shorthand for an immutable class (only edited on creation).
-    public record NodeData(long id, double lat, double lon); //Stores in a Way to hold lat/long for each point in it.
+    public record NodeData(long id, float lat, float lon); //Stores in a Way to hold lat/long for each point in it.
 
-    public record NodeReference(long Id, double lat, double lon, string name, string type); //holds only the node data relevant to the application.
+    public record NodeReference(long Id, float lat, float lon, string name, string type); //holds only the node data relevant to the application.
 
     public record MapDataForJson(long MapDataId, string name, string place, string type, long? WayId, long? NodeId, long? RelationId, int AreaTypeId); //used for serializing MapData, since Geography types do not serialize nicely.
 
-    public record CoordPair(double lat, double lon);
+    public record CoordPair(float lat, float lon);
 
     public record RelationMemberData(long Id, string name, string type);
 }
